sdg popup 추가
